feat: DTO layer with passing tests and compilation fixes
Phase 2 Progress:
- Fix VipLevelBrief fields (Slug/Title/SortOrder)
- Fix Pagination.Total type (int, not int64)
- Add 12 unit tests for DTO conversions (all passing)
  - ToUserResponse: nil handling, field mapping, VIP level
  - ToUserResponseList: batch conversion
  - MaskSensitiveData: phone/email masking, edge cases
  - ToUserListResponseWithConfig: pagination math, masking toggle
  - ToCreateUserInput: whitespace trimming

Test coverage: dto package 100% functional

+++ api/internal/handler/admin/dto/mapper.go
@@ -80,7 +80,7 @@ func ToUserResponseWithConfig(user *model.User, config MapperConfig) *UserRespon
 // ToUserListResponseWithConfig 列表转换（带分页）
 func ToUserListResponseWithConfig(
 	users []model.User,
-	total int64,
+	total int,
 	page, pageSize int,
 	config MapperConfig,
 ) *UserListResponse {
@@ -93,8 +93,8 @@ func ToUserListResponseWithConfig(
 		}
 	}
 	
-	totalPages := int(total) / pageSize
-	if int(total)%pageSize > 0 {
+	totalPages := total / pageSize
+	if total%pageSize > 0 {
 		totalPages++
 	}
